requirements.txt
adal==1.2.7
argcomplete==2.1.2
asttokens==2.4.0
azure-common==1.1.28
azure-core==1.29.4
azure-functions==1.17.0
azure-graphrbac==0.61.1
azure-mgmt-authorization==3.0.0
azure-mgmt-containerregistry==10.1.0
azure-mgmt-core==1.4.0
azure-mgmt-keyvault==10.2.3
azure-mgmt-network==21.0.1
azure-mgmt-resource==22.0.0
azure-mgmt-storage==21.0.0
azureml-core==1.53.0
backcall==0.2.0
backports.tempfile==1.0
backports.weakref==1.0.post1
bcrypt==4.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
comm==0.1.4
contextlib2==21.6.0
contourpy==1.1.0
cryptography==41.0.3
cycler==0.11.0
debugpy==1.7.0
decorator==5.1.1
docker==6.1.3
executing==1.2.0
filelock==3.9.0
fonttools==4.42.1
gitdb==4.0.10
GitPython==3.1.36
humanfriendly==10.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.1
ipykernel==6.25.2
ipython==8.12.2
isodate==0.6.1
jedi==0.19.0
jeepney==0.8.0
Jinja2==3.1.2
jmespath==1.0.1
jsonpickle==3.0.2
jupyter-client==8.3.1
jupyter-core==5.3.1
kiwisolver==1.4.5
knack==0.10.1
MarkupSafe==2.1.2
matplotlib==3.7.3
matplotlib-inline==0.1.6
mpmath==1.2.1
msal==1.24.0
msal-extensions==1.0.0
msrest==0.7.1
msrestazure==0.6.4
ndg-httpsclient==0.5.1
nest-asyncio==1.5.7
networkx==3.0
numpy==1.24.1
oauthlib==3.2.2
opencv-python==4.8.0.76
packaging==23.0
pandas==2.0.3
paramiko==3.3.1
parso==0.8.3
pathspec==0.11.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
pkginfo==1.9.6
platformdirs==3.10.0
portalocker==2.8.2
prompt-toolkit==3.0.39
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyasn1==0.5.0
pycparser==2.21
Pygments==2.16.1
PyJWT==2.8.0
PyNaCl==1.5.0
pyOpenSSL==23.2.0
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.1
requests==2.28.1
requests-oauthlib==1.3.1
scipy==1.10.1
seaborn==0.12.2
SecretStorage==3.3.3
six==1.16.0
smmap==5.0.0
stack-data==0.6.2
sympy==1.11.1
tabulate==0.9.0
thop==0.1.1.post2209072238
torch==2.0.1+cpu
torchaudio==2.0.2+cpu
torchvision==0.15.2+cpu
tornado==6.3.3
tqdm==4.66.1
traitlets==5.9.0
typing-extensions==4.8.0
tzdata==2023.3
ultralytics==8.0.175
urllib3==1.26.13
wcwidth==0.2.6
websocket-client==1.6.3
zipp==3.16.2
